No need to duplicate stuff here and in the show-only-incomplete-todos.md file.
We can expect people to read the Todos guide from beginning to end, so I don't think we need to duplicate the exact text on multiple pages. Hence, I feel that this change makes it look a bit more polished and elegant (even though it can certainly still be improved further).