Tweak tags for blog posts
Blog post tagging had gotten kind of gnarly of late. For one, there was
a "Recent Posts" tag that was applied in a varied number of ways, and
additionally it was unclear how releases and ember-data releases should
be tagged. Consequently there was a variety of junk on the left side of
the page.

To improve the meaning of the categories we have, I've edited tags with
these goals:

* An Ember.js release post should be tagged "Releases". It should not be
  tagged "Recent Posts". It may be tagged "Security".
* An Ember-Data release post should be tagged "Ember-Data". It should
  not be tagged "Recent Posts" or "Releases". It may also be tagged
  "Security"
* A non-release post, or a non-Ember non-Ember-Data release post, should
  be tagged "Recent Posts". Other tags are welcome.

There remain a series of old core team notes on the blog that I have not
re-tagged to "Recent Posts". They seem ok on their own, even though
there is no UI.

I've tweaked the count of displayed recent posts to be much shorter, as
the are not super common.