Fixing the ul/li rendering in Blog posts
The ul/li were not allowed to wrap and their overflow was hidden.
This resulted in content being cut and inaccessible on mobile and
desktop.

This commit tries to fix the issue. It allows `code` content in `li`
to wrap to also prevent overflow on mobile that would cause the annoying
result of "the page scrolls horizontally".

I could not find a `li code` example that was broken by this. If you find
one I'd be happy to find a better solution. In the meantime this allows
for the Ember 3.0 release post to be readable.

:heart: