Fixed Broken Link + Added Emojis
BTW did we remove this section?
The Ember Times is What We Make It 🙌
The Ember Times is a weekly news editorial featuring all the new things that are going on in Emberland. Subscribe to our e-mail newsletter to get the next edition right to your inbox. And if you've always wanted to be an OSS journalist yourself, drop by #topic-embertimes on the Ember Community Slack Chat and write the next edition of the Ember Times together with us!